# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dogly API",
    "description": "Public API V1 (Spec V1 sez. 9). OpenAPI is the mobile client contract.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/dogs/{dog_id}/owner-stories": {
      "get": {
        "tags": [
          "owner-stories"
        ],
        "summary": "List Owner Stories",
        "operationId": "list_owner_stories_v1_dogs__dog_id__owner_stories_get",
        "parameters": [
          {
            "name": "dog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dog Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerStoryListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/dogs/{dog_id}/owner-stories/prepare": {
      "post": {
        "tags": [
          "owner-stories"
        ],
        "summary": "Prepare Owner Story",
        "operationId": "prepare_owner_story_v1_dogs__dog_id__owner_stories_prepare_post",
        "parameters": [
          {
            "name": "dog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dog Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnerStoryPrepareRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerStoryDraftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/dogs/{dog_id}/owner-stories/prepare-audio": {
      "post": {
        "tags": [
          "owner-stories"
        ],
        "summary": "Prepare Owner Story Audio",
        "operationId": "prepare_owner_story_audio_v1_dogs__dog_id__owner_stories_prepare_audio_post",
        "parameters": [
          {
            "name": "dog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dog Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnerStoryAudioPrepareRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerStoryDraftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/dogs/{dog_id}/owner-stories/{draft_id}/confirm": {
      "post": {
        "tags": [
          "owner-stories"
        ],
        "summary": "Confirm Owner Story",
        "operationId": "confirm_owner_story_v1_dogs__dog_id__owner_stories__draft_id__confirm_post",
        "parameters": [
          {
            "name": "dog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dog Id"
            }
          },
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnerStoryConfirmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerStoryConfirmedOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/dogs/{dog_id}/owner-stories/{observation_id}": {
      "patch": {
        "tags": [
          "owner-stories"
        ],
        "summary": "Update Owner Story",
        "operationId": "update_owner_story_v1_dogs__dog_id__owner_stories__observation_id__patch",
        "parameters": [
          {
            "name": "dog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dog Id"
            }
          },
          {
            "name": "observation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Observation Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnerStoryUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerStoryObservationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "owner-stories"
        ],
        "summary": "Delete Owner Story",
        "operationId": "delete_owner_story_v1_dogs__dog_id__owner_stories__observation_id__delete",
        "parameters": [
          {
            "name": "dog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dog Id"
            }
          },
          {
            "name": "observation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Observation Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OwnerReportedFact": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "category": {
            "type": "string",
            "enum": [
              "ROUTINE",
              "PREFERENCE",
              "DIET",
              "HEALTH",
              "GENERAL"
            ],
            "title": "Category"
          },
          "statement": {
            "type": "string",
            "maxLength": 280,
            "minLength": 2,
            "title": "Statement"
          },
          "provenance": {
            "type": "string",
            "const": "OWNER_REPORTED",
            "title": "Provenance",
            "default": "OWNER_REPORTED"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category",
          "statement"
        ],
        "title": "OwnerReportedFact"
      },
      "OwnerStoryAudioPrepareRequest": {
        "properties": {
          "audio_base64": {
            "type": "string",
            "maxLength": 4000000,
            "minLength": 16,
            "title": "Audio Base64"
          },
          "content_type": {
            "type": "string",
            "enum": [
              "audio/m4a",
              "audio/mp4",
              "audio/webm",
              "audio/wav",
              "audio/mpeg"
            ],
            "title": "Content Type"
          }
        },
        "type": "object",
        "required": [
          "audio_base64",
          "content_type"
        ],
        "title": "OwnerStoryAudioPrepareRequest"
      },
      "OwnerStoryConfirmRequest": {
        "properties": {
          "facts": {
            "items": {
              "$ref": "#/components/schemas/OwnerReportedFact"
            },
            "type": "array",
            "maxItems": 8,
            "minItems": 1,
            "title": "Facts"
          }
        },
        "type": "object",
        "required": [
          "facts"
        ],
        "title": "OwnerStoryConfirmRequest"
      },
      "OwnerStoryConfirmedOut": {
        "properties": {
          "observation_id": {
            "type": "string",
            "title": "Observation Id"
          },
          "dog_id": {
            "type": "string",
            "title": "Dog Id"
          },
          "status": {
            "type": "string",
            "const": "CONFIRMED",
            "title": "Status",
            "default": "CONFIRMED"
          },
          "facts": {
            "items": {
              "$ref": "#/components/schemas/OwnerReportedFact"
            },
            "type": "array",
            "title": "Facts"
          }
        },
        "type": "object",
        "required": [
          "observation_id",
          "dog_id",
          "facts"
        ],
        "title": "OwnerStoryConfirmedOut"
      },
      "OwnerStoryDraftOut": {
        "properties": {
          "draft_id": {
            "type": "string",
            "title": "Draft Id"
          },
          "dog_id": {
            "type": "string",
            "title": "Dog Id"
          },
          "transcript": {
            "type": "string",
            "title": "Transcript"
          },
          "facts": {
            "items": {
              "$ref": "#/components/schemas/OwnerReportedFact"
            },
            "type": "array",
            "title": "Facts"
          }
        },
        "type": "object",
        "required": [
          "draft_id",
          "dog_id",
          "transcript",
          "facts"
        ],
        "title": "OwnerStoryDraftOut"
      },
      "OwnerStoryListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OwnerStoryObservationOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "OwnerStoryListOut"
      },
      "OwnerStoryObservationOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "dog_id": {
            "type": "string",
            "title": "Dog Id"
          },
          "facts": {
            "items": {
              "$ref": "#/components/schemas/OwnerReportedFact"
            },
            "type": "array",
            "title": "Facts"
          },
          "confirmed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Confirmed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "dog_id",
          "facts",
          "confirmed_at"
        ],
        "title": "OwnerStoryObservationOut"
      },
      "OwnerStoryPrepareRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 3,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "OwnerStoryPrepareRequest"
      },
      "OwnerStoryUpdateRequest": {
        "properties": {
          "facts": {
            "items": {
              "$ref": "#/components/schemas/OwnerReportedFact"
            },
            "type": "array",
            "maxItems": 8,
            "minItems": 1,
            "title": "Facts"
          }
        },
        "type": "object",
        "required": [
          "facts"
        ],
        "title": "OwnerStoryUpdateRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}